rename server files for quick identification
