+++ src/server/controllers/journal.controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from "express";
 
-import User from "../models/user";
+import User from "../models/user.model";
 import { Entry } from "@server/types";
 
 const getEntries = async (
